refactor(review): searchParamsRef mount-only restore, hook order, move buildSectorFacts to entity lib (PR #563 R4)
