+++ src/entities/sector-signal/lib/sectorFacts.ts
